refactor splashScreen and LogoutButton
